Move ThemeProvider to main App component
