+++ src/pages/Spielstrassen/index.tsx
@@ -1,8 +1,6 @@
 import React, { useEffect } from 'react';
 import { Router, Route, Switch, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
-import { ThemeProvider } from '@material-ui/styles';
-import { createMuiTheme } from '@material-ui/core';
 import { connect } from 'react-redux';
 
 import config from '~/pages/Spielstrassen/config';
@@ -72,16 +70,6 @@ const AppStyles = styled.div`
   }
 `;
 
-const theme = createMuiTheme({
-  palette: {
-    primary: { main: config.colors.interaction },
-    secondary: { main: config.colors.change_4 },
-    error: { main: config.colors.error },
-    info: { main: config.colors.interaction },
-    success: { main: config.colors.label_01 }
-  }
-});
-
 const ScrollToTop = () => {
   const { pathname } = useLocation();
 
@@ -100,34 +88,32 @@ const Spielstrassen = ({ dispatch }) => {
 
   return (
     <AppStyles>
-      <ThemeProvider theme={theme}>
-        <Router history={history}>
-          <ScrollToTop />
-          <Switch>
-            <Route
-              exact
-              path={config.routes.spielstrassen.landing}
-              component={Landing}
-            />
-            <Route
-              exact
-              path={config.routes.spielstrassen.streets}
-              component={Kieze}
-            />
-            <Route
-              exact
-              path={config.routes.spielstrassen.register}
-              component={Register}
-            />
-            <Route
-              exact
-              path={config.routes.spielstrassen.thanks}
-              component={Thanks}
-            />
-            <Route render={() => <Markdown page="nomatch" />} />
-          </Switch>
-        </Router>
-      </ThemeProvider>
+      <Router history={history}>
+        <ScrollToTop />
+        <Switch>
+          <Route
+            exact
+            path={config.routes.spielstrassen.landing}
+            component={Landing}
+          />
+          <Route
+            exact
+            path={config.routes.spielstrassen.streets}
+            component={Kieze}
+          />
+          <Route
+            exact
+            path={config.routes.spielstrassen.register}
+            component={Register}
+          />
+          <Route
+            exact
+            path={config.routes.spielstrassen.thanks}
+            component={Thanks}
+          />
+          <Route render={() => <Markdown page="nomatch" />} />
+        </Switch>
+      </Router>
     </AppStyles>
   );
 };
